# lab4web/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.LoginInput:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: password123
        type: string
    required:
    - email
    - password
    type: object
  handlers.RegisterInput:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: password123
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  handlers.UserResponse:
    properties:
      email:
        example: user@example.com
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@labproject.com
    name: API Support
  description: OAuth2 аутентификация через Яндекс
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Lab Project API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Аутентификация пользователя. Устанавливает access_token и refresh_token
        в cookies.
      parameters:
      - description: Учетные данные
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: Успешный вход
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Неверный формат запроса
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Неверные учетные данные
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties: true
            type: object
      summary: Вход в систему
      tags:
      - Auth
  /auth/logout:
    post:
      description: Отзывает refresh token и очищает cookies
      produces:
      - application/json
      responses:
        "200":
          description: Успешный выход
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Нет refresh token
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Выход из системы
      tags:
      - Auth
  /auth/logout-all:
    post:
      description: Отзывает все refresh токены текущего пользователя
      produces:
      - application/json
      responses:
        "200":
          description: Успешный выход из всех сессий
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Выход из всех сессий
      tags:
      - Auth
  /auth/oauth/{provider}:
    get:
      description: Перенаправляет пользователя на страницу авторизации OAuth2 провайдера
        (Yandex)
      parameters:
      - description: Провайдер OAuth (yandex)
        in: path
        name: provider
        required: true
        type: string
      responses:
        "302":
          description: Перенаправление на OAuth2 провайдера
        "400":
          description: Неподдерживаемый провайдер
          schema:
            additionalProperties: true
            type: object
      summary: Начало OAuth2 аутентификации
      tags:
      - Auth
  /auth/oauth/{provider}/callback:
    get:
      description: Обработка callback от OAuth2 провайдера. Выполняет вход или регистрацию
        пользователя.
      parameters:
      - description: Authorization code
        in: query
        name: code
        required: true
        type: string
      - description: State parameter
        in: query
        name: state
        required: true
        type: string
      responses:
        "200":
          description: Успешная OAuth2 аутентификация
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Неверный запрос
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Неверный state
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties: true
            type: object
      summary: Callback OAuth2 аутентификации
      tags:
      - Auth
  /auth/refresh:
    post:
      description: Обновляет access_token и refresh_token с использованием refresh_token
        из cookie
      produces:
      - application/json
      responses:
        "200":
          description: Токены успешно обновлены
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Нет refresh token
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Неверный refresh token
          schema:
            additionalProperties: true
            type: object
      summary: Обновление токенов
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Создает нового пользователя с указанным email и паролем
      parameters:
      - description: Данные для регистрации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterInput'
      produces:
      - application/json
      responses:
        "201":
          description: Пользователь успешно создан
          schema:
            $ref: '#/definitions/handlers.UserResponse'
        "400":
          description: Неверный формат запроса или пользователь уже существует
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties: true
            type: object
      summary: Регистрация нового пользователя
      tags:
      - Auth
  /auth/whoami:
    get:
      description: Возвращает ID текущего авторизованного пользователя
      produces:
      - application/json
      responses:
        "200":
          description: ID пользователя
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Информация о текущем пользователе
      tags:
      - Auth
securityDefinitions:
  CookieAuth:
    description: JWT токен доступа хранится в HttpOnly cookie. Для авторизации в Swagger
      UI выполните /auth/login, после чего cookie будет отправляться автоматически.
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"